Fix sidebar animation loop, added save icon, and clean up getSavedTartget()
Fixed the issue causing new targets to animate in every time the sidebar is opened/closed.  Renamed the callback in the ‘getSavedTartget’ function to make the function more readable. Converted all button elements in the panel to anchor tags. Also combined the job title & company to one line, and moved the target’s name to the panel header.